al-0.14     * The adaptive locks are fully integrated into the STAMP[1] benchmark suit.     * Transaction nesting is rewritten for performance. Now it reasonably works only if transactions are truly flatten-nested.     * Runtime option to choose lock schemes is removed for best performance.     * GCC's Thread Local Storage (TLS) is used in default for performance.     * Source code is reorganized and copyrighted.

+++ stamp/genome/genome.c
@@ -163,7 +163,7 @@ parseArgs (long argc, char* const argv[])
 		setAdaptMode(1);
 		break;
 	    case 'X':
-		setTranxOvhd(atoi(optarg));
+		setTranxOvhd(atof(optarg));
 		break;
             case '?':
             default:
